backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
pydantic>=2.7
pydantic-settings>=2.2
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
bcrypt==4.1.3
cryptography>=42
pywebpush>=2.0
python-dotenv>=1.0
